requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
langchain
langchain-core
langchain-community
langchain-groq
langgraph
faiss-cpu
sentence-transformers
pypdf
python-multipart